+++ src/CarRental.App/CarRental.Models/Segment.cs
@@ -1,0 +1,10 @@
+﻿namespace CarRental.Models
+{
+    public enum Segment
+    {
+        Unknown,
+        Family,
+        Sport,
+        Premium
+    }
+}
