--- src/CarRental.Models/Reservation.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CarRental.Models
-{
-    public class Reservation
-    {
-        public DateTime StartDate { get; set; }
-
-        public DateTime EndDate { get; set; }
-    }
-}
