--- src/CarRental.Models/Segment.cs
@@ -1,10 +1,0 @@
-﻿namespace CarRental.Models
-{
-    public enum Segment
-    {
-        Unknown,
-        Family,
-        Sport,
-        Premium
-    }
-}
